Ajustando o Form vendas

+++ SistemaVendas/Modelo/Model_Venda.cs
@@ -8,7 +8,7 @@ namespace Modelo
         private DateTime dataVenda;
         private int nFiscal;
         private double total;
-        private double nParcelas;
+        private int nParcelas;
         private string vendaStatus;
         private int idCliente;
         private int idTipoPagamento;
@@ -18,7 +18,7 @@ namespace Modelo
         public DateTime DataVenda { get => dataVenda; set => dataVenda = value; }
         public int NFiscal { get => nFiscal; set => nFiscal = value; }
         public double Total { get => total; set => total = value; }
-        public double NParcelas { get => nParcelas; set => nParcelas = value; }
+        public int NParcelas { get => nParcelas; set => nParcelas = value; }
         public string VendaStatus { get => vendaStatus; set => vendaStatus = value; }
         public int IdCliente { get => idCliente; set => idCliente = value; }
         public int IdTipoPagamento { get => idTipoPagamento; set => idTipoPagamento = value; }
